Add MultiVector.exp()
This uses the Taylor series, which basically applies to anything
that can be linearly combined and also multiplied with itself.
This turns out to have different behavior from how I thought e^V worked
(which makes sense in retrospect): Euler's formula applies only for
multivectors that mimic the imaginary unit.

* Update the docs to reflect the new, corrected behavior.
* Use exp() instead of math.e ** for MultiVector.rotate()
* Define x ** V to be in terms of V.exp() instead:
  x ** V = e ** (V ln x)

These are probably breaking changes, but the old behavior was
demonstrably incorrect, so these are more like bug fixes.